Generic Rasters + SPEEEEDDDDD
Update example to use generic rasters while maintaining some speed.
There are a lot of hacks is in here and no NODATA handling but
the idea is really to see if we could use generic rasters in a
real setting. I think the answer is closer to yes than we may
have thought.